some refactoring of main and functions files

+++ functions/write_file_content.py
@@ -8,30 +8,36 @@ def write_file(working_directory, file_path, content):
     if not target_file.startswith(abs_working_directory):
         return f'Error: Cannot write to "{file_path}" as it is outside the permitted working directory'
     
+    if not os.path.exists(target_file):
+        try:
+            os.makedirs(os.path.dirname(target_file), exist_ok=True)
+        except Exception as err:
+            return f"Error: creating directory: {err}"
+    if os.path.exists(target_file) and os.path.isdir(target_file):
+            return f'Error: "{file_path}" is a directory, not a file'
     try:
-        os.makedirs(os.path.dirname(target_file), exist_ok=True)
-        with open(target_file, "w") as f:
-            print(f"open {target_file}")
+        with open(target_file, "w") as f:            
             f.write(content)          
-        return f'Successfully wrote to "{file_path}" ({len(content)} characters written)'  
+            return f'Successfully wrote to "{file_path}" ({len(content)} characters written)'  
     except Exception as err:
         return f"Error: unable to write to the file: {err}"
 
 
 schema_write_file = types.FunctionDeclaration(
     name="write_file",
-    description="Write or overwrite a file from a specified file path with the specified content, constrained to the working directory.",
+    description="Writes content to a file within the working directory. Creates the file if it doesn't exist.",
     parameters=types.Schema(
         type=types.Type.OBJECT,
         properties={
             "file_path": types.Schema(
                 type=types.Type.STRING,
-                description="The file path to the file to write to, relative to the working directory.",
+                description="Path to the file ti write, relative to the working directory.",
             ),
             "content": types.Schema(
                 type=types.Type.STRING,
-                description="The content to write to the file.",
+                description="Content to write to the file.",
             ),
         },
+        required=["file_path", "content"]
     ),
 )
